fix: CMSG_LEN 返回类型跨平台差异（usize vs u32）
libc crate 的 CMSG_LEN() 在不同目标平台返回 usize 或 u32，
但 cmsg_len 字段总是 usize。减操作需要两端类型一致。
send_fd 侧已有 as usize 转换，recv_fds 侧遗漏。

+++ native/src/zygisk_inject/src/ipc.rs
@@ -169,7 +169,9 @@ pub fn recv_fds(sock: i32) -> alloc::vec::Vec<i32> {
             && (*cmsg).cmsg_level == libc::SOL_SOCKET
             && (*cmsg).cmsg_type == libc::SCM_RIGHTS
         {
-            let payload_len = (*cmsg).cmsg_len - libc::CMSG_LEN(0);
+            // CMSG_LEN() 在 libc crate 中返回 usize 或 u32（取决于目标平台），
+            // 但 cmsg_len 始终是 usize。显式转换为 usize 确保跨平台兼容。
+            let payload_len = (*cmsg).cmsg_len - libc::CMSG_LEN(0) as usize;
             let num_fds = payload_len / core::mem::size_of::<i32>();
             let data_ptr = libc::CMSG_DATA(cmsg) as *const i32;
             for i in 0..num_fds {
